requirements.txt
boto3==1.35.11
botocore==1.35.11
jmespath==1.0.1
pdf2image==1.17.0
pillow==10.4.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
